+++ weblate/accounts/data.py
@@ -27,7 +27,6 @@ DEFAULT_NOTIFICATIONS = [
     (SCOPE_ADMIN, FREQ_INSTANT, "MergeFailureNotification"),
     (SCOPE_ADMIN, FREQ_INSTANT, "ParseErrorNotification"),
     (SCOPE_ADMIN, FREQ_INSTANT, "NewTranslationNotificaton"),
-    (SCOPE_ADMIN, FREQ_INSTANT, "NewComponentNotificaton"),
     (SCOPE_ADMIN, FREQ_INSTANT, "NewAlertNotificaton"),
     (SCOPE_ADMIN, FREQ_INSTANT, "NewAnnouncementNotificaton"),
 ]