+++ weblate/accounts/urls.py
@@ -71,6 +71,7 @@ urlpatterns = [
     url(r"^profile/$", weblate.accounts.views.user_profile, name="profile"),
     url(r"^userdata/$", weblate.accounts.views.userdata, name="userdata"),
     url(r"^unsubscribe/$", weblate.accounts.views.unsubscribe, name="unsubscribe"),
+    url(r"^subscribe/$", weblate.accounts.views.subscribe, name="subscribe"),
     url(r"^watch/(?P<project>[^/]+)/$", weblate.accounts.views.watch, name="watch"),
     url(
         r"^unwatch/(?P<project>[^/]+)/$", weblate.accounts.views.unwatch, name="unwatch"