+++ weblate/formats/base.py
@@ -359,32 +359,32 @@ class TranslationFormat:
         raise NotImplementedError()
 
     @classmethod
-    def get_language_code(cls, code, language_format=None):
+    def get_language_code(cls, code: str, language_format: Optional[str] = None):
         """Do any possible formatting needed for language code."""
         if not language_format:
             language_format = cls.language_format
         return getattr(cls, "get_language_{}".format(language_format))(code)
 
     @staticmethod
-    def get_language_posix(code):
+    def get_language_posix(code: str):
         return code
 
     @staticmethod
-    def get_language_bcp(code):
+    def get_language_bcp(code: str):
         return code.replace("_", "-")
 
     @staticmethod
-    def get_language_posix_long(code):
+    def get_language_posix_long(code: str):
         if code in EXPAND_LANGS:
             return EXPAND_LANGS[code]
         return code
 
     @classmethod
-    def get_language_bcp_long(cls, code):
+    def get_language_bcp_long(cls, code: str):
         return cls.get_language_posix_long(code).replace("_", "-")
 
     @staticmethod
-    def get_language_android(code):
+    def get_language_android(code: str):
         # Android doesn't use Hans/Hant, but rather TW/CN variants
         if code == "zh_Hans":
             return "zh-rCN"
@@ -396,7 +396,7 @@ class TranslationFormat:
         return sanitized.replace("_", "-r")
 
     @classmethod
-    def get_language_java(cls, code):
+    def get_language_java(cls, code: str):
         # Java doesn't use Hans/Hant, but rather TW/CN variants
         if code == "zh_Hans":
             return "zh-CN"
@@ -409,12 +409,14 @@ class TranslationFormat:
         return cls.get_language_bcp(code)
 
     @classmethod
-    def get_language_filename(cls, mask, code):
+    def get_language_filename(
+        cls, mask: str, code: str, language_format: Optional[str] = None
+    ):
         """Return full filename of a language file.
 
         Calculated forfor given path, filemask and language code.
         """
-        return mask.replace("*", cls.get_language_code(code))
+        return mask.replace("*", cls.get_language_code(code, language_format))
 
     @classmethod
     def add_language(cls, filename, language, base):