+++ weblate/formats/exporters.py
@@ -43,27 +43,6 @@ from weblate.utils.site import get_site_url
 # Map to remove control characters except newlines and tabs
 _CHARMAP = dict.fromkeys(x for x in range(32) if x not in (9, 10, 13))
 
-EXPORTERS = {}
-
-
-def register_exporter(exporter):
-    """Register an exporter."""
-    EXPORTERS[exporter.name] = exporter
-    return exporter
-
-
-def get_exporter(name):
-    """Return registered exporter."""
-    return EXPORTERS[name]
-
-
-def list_exporters(translation):
-    return [
-        {"name": x.name, "verbose": x.verbose}
-        for x in sorted(EXPORTERS.values(), key=lambda x: x.name)
-        if x.supports(translation)
-    ]
-
 
 class BaseExporter:
     content_type = "text/plain"
@@ -105,6 +84,10 @@ class BaseExporter:
         if len(plurals) == 1:
             return self.string_filter(plurals[0])
         return multistring([self.string_filter(plural) for plural in plurals])
+
+    @classmethod
+    def get_identifier(cls):
+        return cls.name
 
     def get_storage(self):
         return self.storage_class()
@@ -200,7 +183,6 @@ class BaseExporter:
         return
 
 
-@register_exporter
 class PoExporter(BaseExporter):
     name = "po"
     content_type = "text/x-po"
@@ -240,7 +222,6 @@ class XMLExporter(BaseExporter):
         unit.settarget(word, self.language.code)
 
 
-@register_exporter
 class PoXliffExporter(XMLExporter):
     name = "xliff"
     content_type = "application/x-xliff+xml"
@@ -272,7 +253,6 @@ class PoXliffExporter(XMLExporter):
         return output
 
 
-@register_exporter
 class XliffExporter(PoXliffExporter):
     name = "xliff11"
     content_type = "application/x-xliff+xml"
@@ -282,7 +262,6 @@ class XliffExporter(PoXliffExporter):
     storage_class = xlifffile
 
 
-@register_exporter
 class TBXExporter(XMLExporter):
     name = "tbx"
     content_type = "application/x-tbx"
@@ -291,7 +270,6 @@ class TBXExporter(XMLExporter):
     storage_class = tbxfile
 
 
-@register_exporter
 class TMXExporter(XMLExporter):
     name = "tmx"
     content_type = "application/x-tmx"
@@ -300,7 +278,6 @@ class TMXExporter(XMLExporter):
     storage_class = tmxfile
 
 
-@register_exporter
 class MoExporter(PoExporter):
     name = "mo"
     content_type = "application/x-gettext-catalog"
@@ -360,7 +337,6 @@ class CVSBaseExporter(BaseExporter):
         return self.storage_class(fieldnames=self.fieldnames)
 
 
-@register_exporter
 class CSVExporter(CVSBaseExporter):
     name = "csv"
     content_type = "text/csv"
@@ -380,7 +356,6 @@ class CSVExporter(CVSBaseExporter):
         return text
 
 
-@register_exporter
 class XlsxExporter(CVSBaseExporter):
     name = "xlsx"
     content_type = "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
@@ -406,7 +381,6 @@ class MonolingualExporter(BaseExporter):
         return output
 
 
-@register_exporter
 class JSONExporter(MonolingualExporter):
     storage_class = JsonFile
     name = "json"
@@ -415,7 +389,6 @@ class JSONExporter(MonolingualExporter):
     verbose = _("JSON")
 
 
-@register_exporter
 class AndroidResourceExporter(MonolingualExporter):
     storage_class = AndroidResourceFile
     name = "aresource"
@@ -432,7 +405,6 @@ class AndroidResourceExporter(MonolingualExporter):
         return text.translate(_CHARMAP)
 
 
-@register_exporter
 class StringsExporter(MonolingualExporter):
     storage_class = stringsfile
     name = "strings"