+++ weblate/trans/models/unit.py
@@ -773,7 +773,10 @@ class Unit(models.Model, LoggerMixin):
             else:
                 action = Change.ACTION_CHANGE
         else:
-            action = Change.ACTION_NEW
+            if self.state == STATE_APPROVED:
+                action = Change.ACTION_APPROVE
+            else:
+                action = Change.ACTION_NEW
 
         # Create change object
         Change.objects.create(
@@ -825,6 +828,11 @@ class Unit(models.Model, LoggerMixin):
 
         src = self.get_source_plurals()
         tgt = self.get_target_plurals()
+
+        # Ensure we get a fresh copy of checks
+        # It might be modified meanwhile by propagating to other units
+        if "all_checks" in self.__dict__:
+            del self.__dict__["all_checks"]
 
         old_checks = self.all_checks_names
         create = []